Ignore API version checks in Azurite by default
