+++ src/Aspire.Hosting.Azure.Storage/PublicAPI.Unshipped.txt
@@ -2,3 +2,4 @@
 *REMOVED*static Aspire.Hosting.AzureStorageExtensions.AddAzureStorage(this Aspire.Hosting.IDistributedApplicationBuilder! builder, string! name, System.Action<Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureStorageResource!>!, Aspire.Hosting.ResourceModuleConstruct!, Azure.Provisioning.Storage.StorageAccount!>? configureResource) -> Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureStorageResource!>!
 *REMOVED*Aspire.Hosting.Azure.AzureStorageResource.AzureStorageResource(string! name, System.Action<Aspire.Hosting.ResourceModuleConstruct!>! configureConstruct) -> void
 Aspire.Hosting.Azure.AzureStorageResource.AzureStorageResource(string! name, System.Action<Aspire.Hosting.Azure.AzureResourceInfrastructure!>! configureInfrastructure) -> void
+static Aspire.Hosting.AzureStorageExtensions.WithApiVersionCheck(this Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureStorageEmulatorResource!>! builder, bool enable = true) -> Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.Azure.AzureStorageEmulatorResource!>!
